Refactor modules and tests #minor
Module structure:
 - Detached jfpv1 logic into a separate _jfpv1 module
 - Detached basic validators into a separate _validators module

Tests:
 - Simplified test runner execution with test module loading via
 the parent __init__.py
 - Extended and clarified docstrings for several tests
 - Split existing and new tests into a per-module type structure
 - Added more detailed tests (positive-negative testing) for exceptions
 - Extended 'manual' jfpv1 testing, from raw flattened json output to
 actual fingerprint

CI:
 - Updated test runner execution
