+++ .github/workflows/ci.yml
@@ -32,5 +32,5 @@ jobs:
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       run: |
-        coverage run -m unittest discover -s json_fingerprint -p "test_*.py"
+        coverage run -m unittest json_fingerprint.tests
         coveralls
