+++ json_fingerprint/tests/__init__.py
@@ -1,0 +1,3 @@
+from .test_json_fingerprint import TestJsonFingerprint
+from .test_jfpv1 import TestJfpv1
+from .test_validators import TestValidators
